Hide context local data API.
It's not completely hidden though, it's exposed solely in the internal modules
now. The reason is that the context local data API has inconvenient interaction
with Typeables.

Typeable is required for a data type to be stored as context local data.
However, many Caramia resources have a phantom type that is not known to be
Typeable. I'll try to see if this can be fixed.