Throw away FlextGL. Enter opengl-wrangler.
opengl-wrangler is not yet in hackage and I don't want to make the package
non-hackage installable so for the moment I copied opengl-wrangler generated
modules to the source tree.

This is non-ideal if the user wants to use some raw OpenGL with this library. I
expose some of the key modules to let them get something though.